[refac]: renamed file "io" to "filehandler"
